+++ docs/pages/release_notes.md
@@ -20,10 +20,18 @@ This is a {{ site.pmd.release_type }} release.
 ### New and noteworthy
 
 ### Fixed Issues
+* core
+  * [#3881](https://github.com/pmd/pmd/issues/3881): \[core] SARIF renderer depends on platform default encoding
+  * [#3882](https://github.com/pmd/pmd/pull/3882): \[core] Fix AssertionError about exhaustive switch
+* doc
+  * [#2505](https://github.com/pmd/pmd/issues/2505): \[doc] Improve side bar to show release date
+* java
+  * [#3889](https://github.com/pmd/pmd/pull/3889): \[java] Catch LinkageError in UselessOverridingMethodRule
 
 ### API Changes
 
 ### External Contributions
+* [#3883](https://github.com/pmd/pmd/pull/3883): \[doc] Improve side bar by Adding Release Date - [@jasonqiu98](https://github.com/jasonqiu98)
 
 {% endtocmaker %}